Fix the existence state of bundle.min.css

+++ Bundler/sass-minify.dart
@@ -4,6 +4,7 @@ import 'package:sass/sass.dart' as sass;
 void main(List<String> arguments) {
   var result = sass.compileToResult(arguments[0], style: sass.OutputStyle.expanded, quietDeps: true, verbose: true, sourceMap: false, charset: true);
   var code = new File(arguments[1]);
-  code.deleteSync();
+  if (code.existsSync()) code.deleteSync();
+  code.createSync();
   code.writeAsStringSync(result.css);
 }
